relocated AddTodoForm, InputWithLabel, TodoList, TodoListItem, and all css files to the componenets folder in the src directory. Imported prop-types, defined proptypes of functions as new object
